+++ examples/file_generation_mode/lib/union/to_json/json_key.dart
@@ -21,7 +21,7 @@ abstract class Post {
 }
 
 @Union()
-abstract class GetPostsResponse {
+sealed class GetPostsResponse {
   const GetPostsResponse._();
 
   factory GetPostsResponse.error() = GetPostsResponseError;