+++ package/lib/src/contributors/generators/from_json.dart
@@ -1,12 +1,12 @@
 import 'package:analyzer/dart/element/element.dart';
 import 'package:analyzer/dart/element/type.dart';
 import 'package:data_class_plugin/src/annotations/json_key_internal.dart';
-import 'package:data_class_plugin/src/backend/core/generators/generator.dart';
-import 'package:data_class_plugin/src/common/code_writer.dart';
+import 'package:data_class_plugin/src/common/generator.dart';
 import 'package:data_class_plugin/src/contributors/class/utils.dart';
 import 'package:data_class_plugin/src/extensions/extensions.dart';
 import 'package:data_class_plugin/src/json_key_name_convention.dart';
 import 'package:data_class_plugin/src/options/data_class_plugin_options.dart';
+import 'package:tachyon/tachyon.dart';
 
 class FromJsonGenerator implements Generator {
   /// Shorthand constructor
@@ -198,7 +198,7 @@ class FromJsonGenerator implements Generator {
   }) {
     final String? fieldType = type.element!.name;
 
-    if (type.isDynamic) {
+    if (type is DynamicType) {
       _codeWriter
         ..writeln('// FIXME: variable is dynamic or contains a type that is not yet declared')
         ..writeln('$parentVariableName, ');