+++ package/test/contributors/annotations/data_class/test_files/in_1_simple.dart
@@ -1,0 +1,8 @@
+import 'package:data_class_plugin/data_class_plugin.dart';
+
+@DataClass()
+class User {
+  int get id;
+
+  String get username;
+}