+++ package/test/contributors/annotations/data_class/test_files/out_5_default_value.dart
@@ -1,0 +1,20 @@
+part 'in_5_default_value.gen.dart';
+
+@DataClass()
+abstract class User {
+  User.ctor();
+
+  /// Default constructor
+  factory User({
+    required int id,
+    required String username,
+    String email,
+  }) = _$UserImpl;
+
+  int get id;
+
+  String get username;
+
+  @DefaultValue('')
+  String get email;
+}