+++ package/test/contributors/annotations/union/test_files/in_2_positional_parameter.dart
@@ -1,0 +1,7 @@
+import 'package:data_class_plugin/data_class_plugin.dart';
+
+@Union()
+class AsyncResult {
+  factory AsyncResult.data(int data) = AsyncResultData;
+  factory AsyncResult.error(Object error) = AsyncResultError;
+}