+++ package/test/contributors/annotations/union/test_files/in_3_named_parameter.dart
@@ -1,0 +1,7 @@
+import 'package:data_class_plugin/data_class_plugin.dart';
+
+@Union()
+class AsyncResult {
+  factory AsyncResult.data({required int data}) = AsyncResultData;
+  factory AsyncResult.error({required Object error}) = AsyncResultError;
+}