+++ package/test/contributors/annotations/union/test_files/out_5_to_json.dart
@@ -1,0 +1,14 @@
+part 'in_5_to_json.gen.dart';
+
+@Union(
+  toJson: true,
+)
+sealed class AsyncResult {
+  const AsyncResult._();
+
+  factory AsyncResult.data() = AsyncResultData;
+  factory AsyncResult.error() = AsyncResultError;
+
+  /// Converts [AsyncResult] to [Map] json
+  Map<String, dynamic> toJson();
+}